server/requirements.txt
fastapi 
pygments
uvicorn[standard]
rq
redis